Code additions
DiracKernels added to PorousFlow.  These model the types of line sinks of interest to PorousFlow simulations.
The sinks may be proportional to fluid relative permeability, fluid mobility, enthalpy, internal energy, or the mass fraction of a given species.
The sinks may be piecewise linear functions of porepressure of a phase or temperature.
The sinks may represent wellbores as the Peaceman formulation is implemented.

The PorousFlowSumQuantity UserObject is designed to store the mass or energy flux to/from the DiracKernels.
The PorousFlowPlotQuantity Postprocessor extracts the mass or energy flux from this UserObject.

Fixes #7997
Fixes #7993
Fixes #7992

Also, now the userobject in #8046 is not necessary since i use "nodal" Material instead (as per #8238), so
Fixes #8046