Postprocessors now correctly use lumped mass and energy like Kernels do.
Fixes #8234

Also: j2_hard1_mod test minor regolding.

This is a heavy test and i do not know when it started to differ.  Without this re-gold, the CSV diff is 2.1E-5 on the first nonzero entry of eq_pl_strain, which, in absolute terms, is only a difference of 1E-10 in strain (since that first eq_pl_strain is 2.3E-5).